move messages to another file
